+++ src/components/pages/apps/genes/gsea/gsea-plot/bubble/gsea-bubble-svg.tsx
@@ -1,4 +1,4 @@
-import { useCallback, useMemo, useRef, useState } from 'react'
+import { useMemo } from 'react'
 import { useGseaBubbleSettings } from './gsea-bubble-settings-store'
 
 import { AxisBottomSvg } from '../../../../../../plot/axes/svg-axis'
@@ -20,11 +20,7 @@ import { IPos } from '@/interfaces/pos'
 import { svgPointToScreen } from '@/lib/graphics/svg'
 import { ILim } from '@/lib/math/math'
 import { useSVG } from '@/providers/svg-provider'
-import { createPortal } from 'react-dom'
-import {
-  TOOLTIP_CLEAR_MS,
-  type ITooltip,
-} from '../../../../matcalc/apps/heatmap/heatmap-svg'
+import { useTooltip } from '@/providers/tooltip-provider'
 import { IDisplayAxis } from '../../../../matcalc/apps/volcano/volcano-plot-svg'
 import { IGseaBubble } from '../gsea-plot-store'
 import { IBubblePoint, useGseaBubbleContext } from './gsea-bubble-provider'
@@ -62,6 +58,7 @@ function GseaBubbleLegendSvg() {
   const { plots, globalXLim } = useGseaBubbleContext()
   const { settings } = useGseaBubbleSettings()
   const { settings: edbSettings } = useEdbSettings()
+  const { showTooltip, hideTooltip } = useTooltip()
 
   const sizes = settings.legend.bubbles.sizes
 
@@ -333,55 +330,31 @@ function BubblePlot({
 export function GseaBubblePlotSvg() {
   const { plots, points, xlims } = useGseaBubbleContext()
   const { ref: svgRef } = useSVG()
-  //const containerRef = useRef<HTMLDivElement | null>(null)
+
   const { settings } = useGseaBubbleSettings()
 
-  const tooltipRef = useRef<HTMLDivElement>(null)
+  const { showTooltip, hideTooltip } = useTooltip()
 
-  const [toolTipInfo, setToolTipInfo] = useState<
-    (ITooltip & { plot: IGseaBubble }) | null
-  >(null)
+  function handleVariantEnter(plot: IGseaBubble, row: number, p: IPos) {
+    const screenP = svgPointToScreen(svgRef.current, p)
 
-  const timeoutRef = useRef<ReturnType<typeof setInterval> | null>(null)
-
-  const handleVariantEnter = useCallback(
-    (plot: IGseaBubble, row: number, p: IPos) => {
-      if (timeoutRef.current) {
-        clearTimeout(timeoutRef.current)
-      }
-
-      const screenP = svgPointToScreen(svgRef.current, p)
-
-      //const rect = containerRef.current!.getBoundingClientRect()
-
-      const newP = {
-        x: screenP.x,
-        y: screenP.y,
-      }
-
-      setToolTipInfo({
-        ...toolTipInfo,
-        plot,
-        pos: newP,
-        cell: { row: row, col: 0 },
-      })
-    },
-    []
-  )
-
-  const handleVariantLeave = useCallback(() => {
-    if (timeoutRef.current) {
-      clearTimeout(timeoutRef.current)
+    const newP = {
+      x: screenP.x,
+      y: screenP.y,
     }
 
-    // wait before removing. if we re-enter quickly, the tooltip won't flicker
-    // as this timeout will be cancelled so the tooltip won't disappear
-    // and will be moved to next location
-    timeoutRef.current = setTimeout(
-      () => setToolTipInfo(null),
-      TOOLTIP_CLEAR_MS
-    )
-  }, [])
+    showTooltip({
+      pos: newP,
+      content: (
+        <>
+          <p className="font-semibold">{`${plot.genesets[row]!.name}`}</p>
+          <p>{`${plot.nes.label}: ${plot.genesets[row]!.nes.toFixed(2)}`}</p>
+          <p>{`-log10(${plot.log10q.label}): ${plot.genesets[row]!.log10q.toFixed(2)}`}</p>
+          <p>{`${plot.size.label}: ${plot.genesets[row]!.size}`}</p>
+        </>
+      ),
+    })
+  }
 
   const { svg, width, height } = useMemo(() => {
     //const huedata = hue ? getNumCol(df, findCol(df, hue)) : []
@@ -451,7 +424,7 @@ export function GseaBubblePlotSvg() {
                   innerPlotWidth={innerPlotWidth}
                   innerPlotHeight={innerPlotHeight}
                   handleVariantEnter={handleVariantEnter}
-                  handleVariantLeave={handleVariantLeave}
+                  handleVariantLeave={hideTooltip}
                 />
               </g>
             ))}
@@ -474,28 +447,8 @@ export function GseaBubblePlotSvg() {
   }
 
   return (
-    <>
-      <SvgBase width={width} height={height} scale={settings.page.scale}>
-        {svg}
-      </SvgBase>
-
-      {toolTipInfo &&
-        createPortal(
-          <div
-            ref={tooltipRef}
-            className="fixed z-50 rounded-theme bg-black/60 p-3 text-xs text-white opacity-100"
-            style={{
-              left: toolTipInfo.pos.x,
-              top: toolTipInfo.pos.y,
-            }}
-          >
-            <p className="font-semibold">{`${toolTipInfo.plot.genesets[toolTipInfo.cell.row]!.name}`}</p>
-            <p>{`${toolTipInfo.plot.nes.label}: ${toolTipInfo.plot.genesets[toolTipInfo.cell.row]!.nes.toFixed(2)}`}</p>
-            <p>{`-log10(${toolTipInfo.plot.log10q.label}): ${toolTipInfo.plot.genesets[toolTipInfo.cell.row]!.log10q.toFixed(2)}`}</p>
-            <p>{`${toolTipInfo.plot.size.label}: ${toolTipInfo.plot.genesets[toolTipInfo.cell.row]!.size}`}</p>
-          </div>,
-          document.body
-        )}
-    </>
+    <SvgBase width={width} height={height} scale={settings.page.scale}>
+      {svg}
+    </SvgBase>
   )
 }