+++ src/components/pages/apps/wgs/oncoplot/clinical-tracks-svg.tsx
@@ -1,5 +1,5 @@
+import type { IBlock } from '@/components/pages/apps/matcalc/apps/heatmap/heatmap-settings-store'
 import { YAxis, type Axis } from '@/components/plot/axes/axis'
-import type { IBlock } from '@/components/plot/heatmap/heatmap-svg-props'
 import { SvgLine } from '@/components/plot/svg-line'
 import { SvgRect } from '@/components/plot/svg-rect'
 import { SvgText } from '@/components/plot/svg-text'